refactor: split presets into configs/ directory, add ./node subpath export
