+++ packages/holocron-config/tsdown.config.ts
@@ -1,6 +1,6 @@
 import { defineConfig } from "tsdown";
 export default defineConfig({
-	entry: ["index.ts"],
+	entry: ["index.ts", "configs/*.ts"],
 	format: "esm",
 	fixedExtension: false,
 	dts: true,
